+++ example/src/controllers/Users.ts
@@ -1,0 +1,8 @@
+import { Controller } from 'spooked';
+
+@Controller()
+class Users {
+
+}
+
+export default Users;