+++ src/index.ts
@@ -8,3 +8,4 @@
  */
 
 export { Spooked } from './core/Spooked';
+export { Controller } from './decorators/Controller';