+++ src/types/IController.ts
@@ -1,0 +1,4 @@
+export interface IController {
+    target: Function;
+    name: string;
+}