+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 ## 8/1/2021
+- Add Controller decorator.
+- Add storage manager.
 - Add server manager.
 - Add MySQL example.
 - Add MySQL and MongoDB drivers for connection.